refactor: profile page index -> components
