+++ frontend/app/profilePage/utils.ts
@@ -1,0 +1,6 @@
+export function formatCount(n?: number) {
+  if (n === undefined || n === null) return '—';
+  if (n >= 1000) return `${(n / 1000).toFixed(n % 1000 === 0 ? 0 : 1)}K`;
+  return String(n);
+}
+
