+++ frontend/components/UserDashboard.tsx
@@ -92,7 +92,9 @@ const UserDashboard = ({ user, onSignOut }: Props) => {
 
   return (
     <View>
-      <Text style={{ marginBottom: 10 }}>Welcome, {user.email}</Text>
+      <Text style={{ marginBottom: 10 }}>
+        Welcome, {user?.email || 'Guest'}
+      </Text>
       <View style={styles.buttonContainer}>
         <Button title="Call Protected Backend" onPress={callProtectedBackend} />
         <Button title="Get User Profile" onPress={getUserProfile} />
